Python/Movie_Success_Analysis_1990_2025/requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
jupyter
jupyterlab
ipykernel
ipython
nbformat
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
pytest
pytest-cov
black
flake8
python-dotenv
joblib
statsmodels
pyyaml
openpyxl
tqdm
requests
sqlalchemy